type, name, gender, and stat substitution
- also created a new shared forms file with Stat and AddedPokes (will likely change it later but stat was being referenced from substitution and it needed to move)
